Assembly Programm to create a RGB pattern
The code works and is flashable onto the microcontroller. The RGB LEDs
are programmed to display every color in a consistent sequence, but I
want to change the order to create a more Rainbow like experience. Frequency is a another factor that will be rearranged later on.For now I am satisfied enough for how it went out :D